requirements.txt
pandas
sqlalchemy
psycopg2-binary
alpha_vantage
yfinance
dbt-postgres
